set engine table init flags to 0
When openssl looks up a default engine for RAND_bytes(),
it causes the engine_table_select() routine to be called.
This routine ALWAYS initializes the engine, see:
    https://github.com/openssl/openssl/blob/master/crypto/engine/eng_table.c#L245

SoftHSMv2 was hitting this issue becuase C_Initialize() was invoking a RAND_bytes()
call, and thus calling engine_unlocked_init(), which would call C_Initialize()
and start the whole chain over again.

This was causing make check, as well as other issues.

Set the engine table init flags to NOT call this initialize routine. The initialization
process is described in "Default implementations" section here:
  https://www.openssl.org/docs/man1.1.0/crypto/ENGINE_unregister_ECDSA.html

Signed-off-by: William Roberts <william.c.roberts@intel.com>